Tracked down the error to a difference between a FK getting hashed as a long vs as an int later.  Also some small differences in calculated results.  Now passing on sqlite and postgres!!

+++ ratings/models.py
@@ -26,7 +26,7 @@ class RatedItemBase(models.Model):
         content_field = get_content_object_field(self)
         if is_gfk(content_field):
             uniq = (getattr(self, content_field.ct_field).pk,
-                    getattr(self, content_field.fk_field))
+                    long(getattr(self, content_field.fk_field)))
         else:
             uniq = getattr(self, content_field.name).pk
         hashed = sha_constructor(str(uniq)).hexdigest()
